fix: Fix UserTweets (not yet)
