+++ src/apis/tweet.js
@@ -17,15 +17,5 @@ export default {
     return apiHelper.get(`/tweets/${tweetId}`, {
       headers: { Authorization: `Bearer ${getToken()}` }
     })
-  },
-  addLike ({ tweetId }) {
-    return apiHelper.post(`/tweets/${tweetId}/like`, null, {
-      headers: { Authorization: `Bearer ${getToken()}` }
-    })
-  },
-  deleteLike ({ tweetId }) {
-    return apiHelper.post(`/tweets/${tweetId}/unlike`, null, {
-      headers: { Authorization: `Bearer ${getToken()}` }
-    })
   }
 }
