+++ src/components/UserTweets.vue
@@ -3,7 +3,7 @@
     <!-- test -->
     <Spinner v-if="isLoading"/>
     <!-- 原先在此渲染，改在父層渲染 -->
-    <!-- <div v-for="tweet in tweets" :key="tweet.id" class="user-tweets"> -->
+    <div v-for="tweet in tweets" :key="tweet.id" class="user-tweets">
       <!--  到當前使用者的路由 -->
       <!-- <router-link
         :to="{ name: 'user', params: { id: tweet.User.id } }"
@@ -30,13 +30,6 @@
           @click.stop.prevent="handleReplyModal(tweet)"
           class="icon-section">
             <div class="footer-icon reply-icon d-flex"></div>
-            <!-- todo: addReply -->
-            <!-- 版壞掉 -->
-            <!-- <img
-                class="counter reply-icon"
-                src="./../assets/pictures/reply.png"
-                alt=""
-              /> -->
             <span 
             class="counter reply-count">{{ tweet.replyCounts }}
             </span>
@@ -67,9 +60,12 @@
           </div>
         </div>
       </div>
-    <!-- </div> -->
+    </div>
     <!-- todo: not yet -->
-    <ReplyModal
+    <!-- 改 -->
+    <ReplyModal 
+      v-show="show === true"
+      :show="show === true"
       :tweet="tweetActive"
       @after-reply-submit="afterReplySubmit"
     />
@@ -82,7 +78,8 @@ import Spinner from '../components/Spinner.vue'
 import ReplyModal from "./../components/ReplyModal.vue"
 import { fromNowFilter } from "./../utils/mixins";
 
-import tweetsAPI from "./../apis/tweet";
+// import likesAPI from "./../apis/like";
+import userAPI from "./../apis/user";
 import { Toast } from "../utils/helpers";
 // 載入 Vuex
 import { mapState } from "vuex";
@@ -94,66 +91,108 @@ export default {
     ReplyModal,
     Spinner
   },
-  props: {
-    initialTweets: {
-      type: Array,
-      required: true,
-    },
-  },
   data() {
     return {
-      // userTweets: [],
+      show: false,
       tweetActive: {},
       isLoading: true,
-      tweets: this.initialTweets,
+      tweets: [],
     };
   },
   // 從 Vuex 取得 currentUser 的資料
   computed: {
     ...mapState(["currentUser"]),
   },
+  created() {
+    // 取得動態路由位置
+    const { id: userId } = this.$route.params;
+    this.fetchUserTweets(userId);
+  },
+  // 追蹤路由變化
+  beforeRouteUpdate(to, from, next) {
+    const { id } = to.params;
+    this.fetchUserTweets(id);
+    next();
+  },
   
   methods: {
-    
-    // todo: 有問題
-    async addLike (tweetId) {
+    // OK 向資料庫撈取 使用者 tweets 資料
+    async fetchUserTweets(userId) {
       try {
-        const { data } = await tweetsAPI.addLike({ tweetId })
+        const response  = await userAPI.getUserTweets({ userId });
+        
+        if (response.status !== 200) {
+          throw new Error(response.message);
+        }
+        const tweetsData = response.data
+        
+        this.tweets = tweetsData
 
-        if (data.status !== 'success') {
-          throw new Error(data.message)
-        }
-        this.tweet.isLiked = true
-        this.tweet.likesCounts += 1
+        this.isLoading = false;
       } catch (error) {
+        this.isLoading = false;
         Toast.fire({
-          icon: 'error',
-          title: '無法按喜歡，請稍後再試'
-        })
+          icon: "error",
+          title: "無法取得 tweets",
+        });
       }
     },
+    // 未串 API 前測試
+    // addLike(tweetId) {
+    //   console.log('tweetId', tweetId)
+    // },
+    // deleteLike(tweetId) {
+    //   console.log('tweetId', tweetId)
+    // },
     // todo: 有問題
-    async deleteLike (tweetId) {
-      try {
-        const { data } = await tweetsAPI.deleteLike({ tweetId })
-        if (data.status !== 'success') {
-          throw new Error(data.message)
-        }
-        this.tweet.isLiked = false
-        this.tweet.likesCounts -= 1
-      } catch (error) {
-        Toast.fire({
-          icon: 'error',
-          title: '無法按喜歡，請稍後再試'
-        })
-      }
-    },
+    // async addLike (tweetId) {
+    //   try {
+    //     const response = await likesAPI.likeTweet(tweetId);
+        
+    //     // 錯誤狀態處理
+    //       if (response.data.status === "error") {
+    //         throw new Error(response.data.message);
+    //       }
+    //       // 更新 like 數
+          
+    //     // if (data.status !== 'success') {
+    //     //   throw new Error(data.message)
+    //     // }
+    //     // this.tweet.isLiked = true
+    //     // this.tweet.likesCounts += 1
+    //   } catch (error) {
+    //     Toast.fire({
+    //       icon: 'error',
+    //       title: '無法按喜歡，請稍後再試'
+    //     })
+    //   }
+    // },
+    // todo: 有問題
+    // async deleteLike (tweetId) {
+    //   try {
+    //     const  data  = await likesAPI.unLikeTweet({ tweetId })
+    //     console.log('delete like', data)
+
+    //     // if (data.status !== 'success') {
+    //     //   throw new Error(data.message)
+    //     // }
+    //     // this.tweet.isLiked = false
+    //     // this.tweet.likesCounts -= 1
+    //   } catch (error) {
+    //     Toast.fire({
+    //       icon: 'error',
+    //       title: '無法按喜歡，請稍後再試'
+    //     })
+    //   }
+    // },
     // 
     handleReplyModal(tweet) {
-      // 取出整包內容儲存到 data
+      this.show = true;
       this.tweetActive = tweet;
-      console.log('tweet', tweet)
     },
+    // 須要 call API ?
+  
+    
     // todo: 未完成
     afterReplySubmit(payload) {
       console.log('payload',payload)
